validation for existing email

+++ JavaScript/Lab/Lab 11/app.js
@@ -6,6 +6,11 @@ function signup() {
     if (!email || !password) return;
 
     let newUser = { email, password };
+    // let newUser = {
+    //     email: email,
+    //     password: password
+    // };
+
     saveInDB(newUser)
 }
 
@@ -13,10 +18,24 @@ function signup() {
 function saveInDB(newUser) {
     let users = JSON.parse(localStorage.getItem('users')) || [];
     console.log(users);
+    let isAlreadyExisting = checkUserExistance(users, newUser);
+    if (isAlreadyExisting) return;
 
     users.push(newUser);
 
     users = JSON.stringify(users);
-        
+
     localStorage.setItem("users", users);
+}
+
+
+function checkUserExistance(users, newUser) {
+    let isAlreadyExisting = false;
+    for (let i = 0; i < users.length; i++) {
+        if (users[i].email === newUser.email) {
+            isAlreadyExisting = true;
+            break;
+        }
+    }
+    return isAlreadyExisting;
 }
